Complete Neuron definition
This patch completes the neuron's API with the following changes:

Neuron's constructor changes
- Add empty constructor
- Add copy constructor
- Add an empty destructor

Neuron's getters changes
- Add need_refresh method
- Add factors_size method
- Add factors method
- Add factor_changes method
- Add last_factor_changes method
- Add delta method
- Add before_layer method

Neuron's setters changes
- Add enable_bias method
- Add disable_bias method
- Add set_delta method
- Add set_output_refresh method
- Add unset_output_refresh method

Other changes:
- Set before_layer pointer as a shared_ptr (smart pointer)
- Refactor all constructors
- Refactor output method

Notice that, since now, the output don't need to be refreshed because it
will be done automatically. That implies that output method can not be a
const method but ensures that the correct output is always used.