Reorganized code
With test files, I felt that the tree will be a mess in a few weeks, so
I reorganize all the code in the following structure:

BASEDIR
|
-- bin (here must be the all binaries)
|
-- obj (all c++ objects must be here)
|
-- src (all app's code must be here)
|
-- test (test files must be here)

So the idea is that the BASEDIR must contain 4 directories where src and
test are designed to containt the project's code organized, wheareas the
bin and obj must be created on demand by the makefile.

BASEDIR can contain other stuff like the LICENSE or README.MD and of
course the main makefile.