Refactor neuron's structure
I redo a lot of the neuron structure, to do it easier to use. The
changes are:

- No more templates.
- Rename classes and namespaces to use lower case names
- Rename SigmoidNeuron into sigmoid (it is easy to understand that it is
  a neuron because it is in mp::neuron::sigmoid)
- Remove pointers to the before layer.
- Add two protected methods that calculate the neuron output. These
  methods are virtual and must be redefined in the derived classes.
- One of the methods that calculate the neuron output receives a
  reference to a vector of doubles, and it will be used to get the real
  output. The another one receives a reference to a vector of
  mp::neuron::base pointers.
- Add two new methods called refresh, that call to the virtual calculate
  output and update the internal output variable.

With these changes, I can archieve the same results that I had with the
templates, smart pointers and type traits before, but the neuron's
structure is simpler and easier to manage.

Well, there is one difference, before I do this, the neuron had the
ability to auto update themselves, and now the network will have to do
it. That was archieved with the pointers to the before layer, if I find
a way to recover this, I will write it, but until them, the neurons must
be refreshed to update it's outputs.