Created a basic network class
The network is not capable to do any useful yet, but it can be
constructed in a valid state and it can be fed.

Also you can retreive a (smart) pointer of a neuron, and update it.

Also, neuron's refresh and calculate_output methods receive now
a smart pointer's vector instead of a base *.